Rename shared dir
